Three question Setup
Issue: The 'Three' above is hard-coded.
requirement: The user should be prompted with only that many questions that are not setup.
For example if there is already one question setup. but later the requirement comes from application
that 3 questions are required from now on. The user should only be prompted for 2 more questions, since
one has already been setup.